✨ Make application options available to middleware.
We need to be able to reflect on the application structure in order
see options that are available to it. This is what we need to do
things like make a sitemap plugin that can reason about all routes
together.

This stores the options, and then makes them available to anybody
interested in the on the middleware stack.