✨ respond with redirects
One of the nice things about using continuations is that you can drop
the current frame of execution and just resume where you were before.
This lets you do that with HTTP responses. So, for example, if you
need to redirect, then it will drop what you're doing and perform the
redirect, doing any necessary unwinding as you go along.